edit to ref

+++ Entities/Models/Contact.cs
@@ -13,7 +13,7 @@ namespace Entities.Models
         [Key]
         public int Id { get; set; }
         public Guid? UserId { get; set; }
-        public uint? Telephone { get; set; }
+        public long? Telephone { get; set; }
 
         // Relational
         //[ForeignKey(nameof(UserId))]
